move map and node to different package
