+++ src/main/java/edu/wpi/teamc/map/Map.java
@@ -1,4 +1,4 @@
-package edu.wpi.teamc.navigation;
+package edu.wpi.teamc.map;
 
 import java.util.HashMap;
 import java.util.LinkedList;
